Version 2.0
Rewrite to include CLI app and YAML files

Once again include the CLI app in this package, so that only one package
needs to be installed.

Specify manifests in YAML files (because they're [easier to read][yaml]
than JSON files), and only allow a single product in a YAML file.

[yaml]: https://burgers.io/yaml-configuration