Change the structure of the schema file
Apparently the object that js-yaml creates has its keys in alphabetical
order, not a top-to-bottom order of keys as defined by the YAML file.
(OK, maybe it was silly to depend on undefined key-order semantics in
the first place.)

Because order is so important, the versions key is now an array rather
than an object, so that we can get the defined order of the versions.